+++ packages/desktop/app/javascripts/Main/FileBackups/FileBackupsManager.ts
@@ -1,4 +1,10 @@
-import { FileBackupRecord, FileBackupsDevice, FileBackupsMapping } from '@web/Application/Device/DesktopSnjsExports'
+import {
+  FileBackupRecord,
+  FileBackupsDevice,
+  FileBackupsMapping,
+  FileBackupReadToken,
+  FileBackupReadChunkResponse,
+} from '@web/Application/Device/DesktopSnjsExports'
 import { AppState } from 'app/AppState'
 import { shell } from 'electron'
 import { StoreKeys } from '../Store/StoreKeys'
@@ -6,13 +12,14 @@ import path from 'path'
 import {
   deleteFile,
   ensureDirectoryExists,
-  moveFiles,
+  moveDirContents,
   openDirectoryPicker,
   readJSONFile,
   writeFile,
   writeJSONFile,
 } from '../Utils/FileUtils'
 import { FileDownloader } from './FileDownloader'
+import { FileReadOperation } from './FileReadOperation'
 
 export const FileBackupsConstantsV1 = {
   Version: '1.0.0',
@@ -21,6 +28,8 @@ export const FileBackupsConstantsV1 = {
 }
 
 export class FilesBackupManager implements FileBackupsDevice {
+  private readOperations: Map<string, FileReadOperation> = new Map()
+
   constructor(private appState: AppState) {}
 
   public isFilesBackupsEnabled(): Promise<boolean> {
@@ -78,8 +87,11 @@ export class FilesBackupManager implements FileBackupsDevice {
     }
 
     const entries = Object.values(mapping.files)
-    const itemFolders = entries.map((entry) => path.join(oldPath, entry.relativePath))
-    await moveFiles(itemFolders, newPath)
+    for (const entry of entries) {
+      const sourcePath = path.join(oldPath, entry.relativePath)
+      const destinationPath = path.join(newPath, entry.relativePath)
+      await moveDirContents(sourcePath, destinationPath)
+    }
 
     for (const entry of entries) {
       entry.absolutePath = path.join(newPath, entry.relativePath)
@@ -184,6 +196,30 @@ export class FilesBackupManager implements FileBackupsDevice {
       }
 
       await this.saveFilesBackupsMappingFile(mapping)
+    }
+
+    return result
+  }
+
+  async getFileBackupReadToken(record: FileBackupRecord): Promise<FileBackupReadToken> {
+    const operation = new FileReadOperation(record)
+
+    this.readOperations.set(operation.token, operation)
+
+    return operation.token
+  }
+
+  async readNextChunk(token: string): Promise<FileBackupReadChunkResponse> {
+    const operation = this.readOperations.get(token)
+
+    if (!operation) {
+      return Promise.reject(new Error('Invalid token'))
+    }
+
+    const result = await operation.readNextChunk()
+
+    if (result.isLast) {
+      this.readOperations.delete(token)
     }
 
     return result