+++ packages/web/src/javascripts/Utils/DateUtils.spec.ts
@@ -1,4 +1,4 @@
-import { addDays } from '@/Utils/DateUtils'
+import { addDaysToDate } from '@standardnotes/utils'
 import { numDaysBetweenDates } from './DateUtils'
 
 describe('date utils', () => {
@@ -6,16 +6,16 @@ describe('date utils', () => {
     it('should return full days diff accurately', () => {
       const today = new Date()
 
-      expect(numDaysBetweenDates(today, addDays(today, 1))).toEqual(1)
-      expect(numDaysBetweenDates(today, addDays(today, 2))).toEqual(2)
-      expect(numDaysBetweenDates(today, addDays(today, 3))).toEqual(3)
+      expect(numDaysBetweenDates(today, addDaysToDate(today, 1))).toEqual(1)
+      expect(numDaysBetweenDates(today, addDaysToDate(today, 2))).toEqual(2)
+      expect(numDaysBetweenDates(today, addDaysToDate(today, 3))).toEqual(3)
     })
 
     it('should return absolute value of difference', () => {
       const today = new Date()
 
-      expect(numDaysBetweenDates(today, addDays(today, 3))).toEqual(3)
-      expect(numDaysBetweenDates(addDays(today, 3), today)).toEqual(3)
+      expect(numDaysBetweenDates(today, addDaysToDate(today, 3))).toEqual(3)
+      expect(numDaysBetweenDates(addDaysToDate(today, 3), today)).toEqual(3)
     })
 
     it('should return 1 day difference between two dates on different days but 1 hour apart', () => {