+++ packages/files/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.14.0](https://github.com/standardnotes/app/compare/@standardnotes/files@1.13.5...@standardnotes/files@1.14.0) (2022-12-01)
+
+### Features
+
+* download and preview files from local backups automatically, if a local backup is available ([#2076](https://github.com/standardnotes/app/issues/2076)) ([28e43d3](https://github.com/standardnotes/app/commit/28e43d37c00c5cd97129c4b83266f7a234bd9857))
+
+## [1.13.5](https://github.com/standardnotes/app/compare/@standardnotes/files@1.13.4...@standardnotes/files@1.13.5) (2022-12-01)
+
+**Note:** Version bump only for package @standardnotes/files
+
 ## [1.13.4](https://github.com/standardnotes/app/compare/@standardnotes/files@1.13.3...@standardnotes/files@1.13.4) (2022-11-28)
 
 **Note:** Version bump only for package @standardnotes/files